Refactor : type refactor
