+++ components/Home/MainBanner.tsx
@@ -1,45 +1,6 @@
 import Image from 'next/image';
-import React, { useEffect, useState } from 'react';
-import styled from 'styled-components';
-
-type mainBannerProps = {
-  $isClick: boolean;
-};
-
-const MainBannerContainer = styled.div<mainBannerProps>`
-  width: 100%;
-  height: ${(props) => (props.$isClick ? '250px' : 'calc(100vh - 52px)')};
-  transition: all 0.8s ease-in-out;
-  position: relative;
-  p {
-    width: 90%;
-    word-wrap: break-word;
-    position: absolute;
-    bottom: 120px;
-    left: 50%;
-    transform: translateX(-50%);
-    font-size: 20px;
-    font-weight: 700;
-    line-height: 28px;
-    color: #fff;
-  }
-  img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    transition: all 0.8s ease-in-out;
-    opacity: ${(props) => (props.$isClick ? '0' : '1')};
-  }
-  @media screen and (min-width: 768px) {
-    height: ${(props) => (props.$isClick ? '400px' : '100vh')};
-    p {
-      bottom: 225px;
-      max-width: 768px;
-      font-size: 28px;
-      line-height: 36px;
-    }
-  }
-`;
+import React from 'react';
+import { MainBannerContainer } from '@/styles/home';
 
 const MainBanner = ({ data, $isClick }: any) => {
   return (
