+++ pages/index.tsx
@@ -1,47 +1,24 @@
 import React, { useState } from 'react';
-import MainBannerSlider from '@/components/Home/MainBannerSlider';
+import MainBannerSlider from '@/components/Home/InfiniteCarousel';
 import HashtagNavBar from '@/components/Home/HashtagNavBar';
 import HashtagContents from '@/components/Home/HashtagContents';
-import Main from '@/components/Home/Main';
-import PopularContents from '@/components/Home/PopularContents';
+import Main from '@/components/Home/HomeMain';
 import MainBanner from '@/components/Home/MainBanner';
 import Header from '@/components/common/Header';
 import Footer from '@/components/common/Footer';
 import PopularContentsCarousel from '@/components/Home/PopularContentsCarousel';
-import styled from 'styled-components';
-import {
-  sliderImages,
-  SliderImage,
-  hashtagImages,
-  hashtagContentsImages,
-  popularContentsImages,
-} from '../public/data';
+import { sliderImages, hashtagImages } from '../public/data';
 import { GetStaticProps, NextPage, GetServerSideProps } from 'next';
 import { useQuery } from 'react-query';
 import axios from 'axios';
-import MobileBottomNav from '@/components/common/MobileBottomNav';
 import { getAuthToken } from '@/api/user';
 import { AUTH_USER, UserResponse } from '@/redux/reducers/userSlice';
 import { useDispatch } from 'react-redux';
 import { setCookie } from 'nookies';
-import api from '../axios/api';
-interface Props {
-  data: SliderImage[];
-  hashtagData: SliderImage[];
-  hashContent: any;
-  popularContent: any;
-}
+import { HomePageLayout } from '@/styles/home';
+import { MainPageProps } from '@/types/home';
 
-const HomePageLayout = styled.div`
-  width: 100%;
-  height: 100vh;
-  overflow: hidden;
-  /* @media screen and (min-width: 768px) {
-    overflow: initial;
-  } */
-`;
-
-const Home: NextPage<Props> = ({
+const Home: NextPage<MainPageProps> = ({
   data,
   hashtagData,
   hashContent,
