+++ styles/imageContainer.ts
@@ -1,4 +1,5 @@
 import styled from 'styled-components';
+import Link from 'next/link';
 
 export const ImageContentsContainer = styled.div`
   max-width: 1200px;
@@ -15,5 +16,30 @@ export const ImageContentsContainer = styled.div`
   }
   @media screen and (max-width: 768px) {
     grid-template-columns: repeat(2, 1fr);
+  }
+`;
+
+export const ImageContentLayout = styled.div`
+  width: 100%;
+  position: relative;
+`;
+
+export const ImageBox = styled(Link)`
+  display: block;
+  position: relative;
+  width: 100%;
+  padding-bottom: 100%;
+  img {
+    object-fit: cover;
+    -webkit-user-drag: none;
+    -khtml-user-drag: none;
+    -moz-user-drag: none;
+    -o-user-drag: none;
+    user-drag: none;
+    -webkit-user-select: none;
+    -khtml-user-select: none;
+    -moz-user-select: none;
+    -o-user-select: none;
+    user-select: none;
   }
 `;
