+++ styles/modal.ts
@@ -6,11 +6,11 @@ export const ModalBackground = styled.div`
   left: 0;
   width: 100%;
   height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
+  /* background-color: rgba(0, 0, 0, 0.5); */
   display: flex;
   align-items: center;
   justify-content: center;
-  z-index: 20;
+  z-index: 999;
 `;
 
 export const ModalContainer = styled.div`
