+++ types/common.d.ts
@@ -1,0 +1,3 @@
+export interface childrenProps {
+  children: ReactNode;
+}
